chore: Remove Scorecard Analysis (#357)
# Pull Request

## Description

This pull request includes changes to the
`.github/workflows/code-checks.yml` file, specifically focusing on the
removal of the Scorecard Analysis job. The most important change is the
removal of the Scorecard Analysis job and its associated steps.

Workflow changes:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L237-L259):
Removed the `run-scorecard-analysis` job, which included steps for
checking out code, running the Scorecard analysis, and uploading results
to code scanning.

Fixes #356