refactor: Move Repositories Page to Index (#379)
# Pull Request

## Description

This pull request includes significant changes to the `dashboard`
application, mainly focusing on the restructuring of repository data
handling and the removal of certain UI components. The most important
changes include the addition of test data, the restructuring of the
repositories page, and the removal of the `HomePageCards` component.

### Data Handling and Testing:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R223-R224):
Added a step to copy test data for repository statistics to the
dashboard data directory.
*
[`test_data/repository_statistics.json`](diffhunk://#diff-49f00b9fd64fb22feb5139f0031f039f5c991cbd82a8605050834651bf9e62daR1-R147):
Added comprehensive test data for various repositories, including
details about commits and languages used.

### UI Changes:

*
[`dashboard/app/(pages)/repositories/page.tsx`](diffhunk://#diff-08608cf160258ca75586d00e5be82714b21d0e205e837646183ef8a8a3971690L1-L31):
Removed the old repositories page implementation.
*
[`dashboard/app/page.tsx`](diffhunk://#diff-e1b87d8e32d6a6029b017f01e52506c46844e0d2a3a7fbd57decc879e5630d6fL1-R28):
Updated the main page to display repository statistics instead of GitHub
stats, incorporating a new layout and linking to individual
repositories.
*
[`dashboard/components/ui/HomePageCards.tsx`](diffhunk://#diff-82fb14677067d322c7e57befde94ca257473572ba7fecba2868805a223657404L1-L52):
Removed the `HomePageCards` component, which previously displayed links
to user and repository details.

### Testing Adjustments:

*
[`dashboard/app/__tests__/page.spec.tsx`](diffhunk://#diff-694f8c76deb82890cdb08dcec7b90eb1575d51cf79f15862257994d5081921e7L10-R10):
Updated the test for the home page to reflect the new heading text
"Repositories".