refactor: Restructure Page Header (#380)
# Pull Request

## Description

This pull request includes changes to the layout and navigation
components of the dashboard application. The most important changes
involve the removal of the repositories link, updating the layout
structure, and renaming a navigation item.

Layout updates:

*
[`dashboard/app/(pages)/layout.tsx`](diffhunk://#diff-eccb9739626633732856b964129e6b2fe77b678efa9c22e404b1576476566583L1-L24):
Removed the `RootLayout` component and its associated metadata, as well
as the import statements for `Metadata`, `React`, `Navbar`, and global
styles.
*
[`dashboard/app/layout.tsx`](diffhunk://#diff-8aae3bf596c8945c84b1f07ce8810f1d9d848a80a9dee7f06d9ba0efaf02b432R5-R6):
Re-added the `Navbar` component to the layout and adjusted the body
structure to include the `Navbar`.
[[1]](diffhunk://#diff-8aae3bf596c8945c84b1f07ce8810f1d9d848a80a9dee7f06d9ba0efaf02b432R5-R6)
[[2]](diffhunk://#diff-8aae3bf596c8945c84b1f07ce8810f1d9d848a80a9dee7f06d9ba0efaf02b432R35)

Navigation updates:

*
[`dashboard/components/NavigationMenuItems.tsx`](diffhunk://#diff-13122630c36e08771ee2250c0cc28c734ef82244685c63fc9255f3171b01c0c8L1-R1):
Removed the `RowsIcon` import and the link to the repositories page.
Renamed the "Overview" link to "Summary".
[[1]](diffhunk://#diff-13122630c36e08771ee2250c0cc28c734ef82244685c63fc9255f3171b01c0c8L1-R1)
[[2]](diffhunk://#diff-13122630c36e08771ee2250c0cc28c734ef82244685c63fc9255f3171b01c0c8L21-L36)