build: Sort Repositories List Alphabetically (#389)
# Pull Request

## Description

This pull request includes changes to the
`dashboard/components/SideBar.tsx` file to improve the user experience
by sorting the repositories in the sidebar menu. The most important
change is the addition of sorting functionality for the repositories.

Improvements to user experience:

*
[`dashboard/components/SideBar.tsx`](diffhunk://#diff-789a55c11f3c7300c5678f37f83847350e681afdd7621d5141a72296b473dd0cR71-R77):
Added sorting functionality to display repositories in alphabetical
order in the sidebar menu.

Fixes #386