refactor: Refactor Zizmor Action (#426)
# Pull Request

## Description

This pull request includes updates to the GitHub Actions workflow,
dependency management, and test setup in the project. The most important
changes involve upgrading the environment, setting up dependencies, and
updating the development dependencies.

GitHub Actions workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L243-R243):
Updated the runner to use `ubuntu-22.04` instead of `ubuntu-latest` and
replaced the setup for `uv` with a custom action for setting up test
dependencies.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L243-R243)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L252-R255)

Dependency management updates:

*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL14-R14):
Updated the `dev` optional dependencies to include `ruff==0.9.2` and
`zizmor==1.2.2`.

Test setup updates:

*
[`tests/tests.just`](diffhunk://#diff-83e2fb4528694cfa62c213eefec53e6aeee0353030bf91ad4f830e69854b7918L9-L12):
Removed the `install-only-root` target and corrected a typo in the
comment for Ruff linting and formatting.
[[1]](diffhunk://#diff-83e2fb4528694cfa62c213eefec53e6aeee0353030bf91ad4f830e69854b7918L9-L12)
[[2]](diffhunk://#diff-83e2fb4528694cfa62c213eefec53e6aeee0353030bf91ad4f830e69854b7918L44-R40)