chore: Fix UI Tests (#443)
# Pull Request

## Description

This pull request includes several updates to dependencies and workflows
in the repository. The most important changes involve updating the
versions of various GitHub Actions and dependencies to ensure
compatibility and leverage the latest features and fixes.

### Dependency Updates

*
[`.github/actions/setup-dashboard-dependencies/action.yml`](diffhunk://#diff-f9870bea14450f1c76432e388b314c2cca7f7f040812d75faab93e56630e1865L8-R8):
Updated `actions/setup-node` from `v4.1.0` to `v4.2.0` to use the latest
version.
*
[`.github/actions/setup-test-dependencies/action.yml`](diffhunk://#diff-ee1e9a84357cfd23d067400706d92b6bb9d8fd8744e81563d9858e8c344df144L10-R12):
Updated `browser-actions/setup-chrome` from `v1.7.2` to `v1.7.3` and
`actions/setup-python` from `v5.3.0` to `v5.4.0` to ensure compatibility
with newer versions.
[[1]](diffhunk://#diff-ee1e9a84357cfd23d067400706d92b6bb9d8fd8744e81563d9858e8c344df144L10-R12)
[[2]](diffhunk://#diff-ee1e9a84357cfd23d067400706d92b6bb9d8fd8744e81563d9858e8c344df144L23-R23)

### Workflow Improvements

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L75-R75):
Updated `github/codeql-action` versions for `upload-sarif`, `init`, and
`analyze` from `v3.28.1` to `v3.28.6` for improved security and
performance.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L75-R75)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L105-R105)
[[3]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L172-R177)
[[4]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L261-R253)
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L190-R191):
Replaced multiple steps for setting up Node.js and Just with a single
step that uses the `setup-dashboard-dependencies` action for
simplification.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L245-R237):
Changed the runner for `run-zizmor` from `ubuntu-22.04` to
`ubuntu-latest` to always use the latest stable version.

### Dependency Configuration

*
[`dashboard/package-lock.json`](diffhunk://#diff-ed953a47c378f3fc19c8e0152346ac0ffa9162d0ffacdbae96c4d9ffbcf479a6L53-R54):
Fixed the versions of `prettier` and `prettier-plugin-tailwindcss` to
specific versions instead of using caret ranges for more predictable
builds.