chore: Reformat Application Colours (#451)
# Pull Request

## Description

This pull request includes several changes to the `dashboard`
application, focusing on updating the global styles, modifying the
`LanguagesBarChart` component, and enhancing the `SideBar` component.
The most important changes include updating the color scheme in the
global styles, changing the layout and labels of the
`LanguagesBarChart`, and adding a variant to the `SideBar`.

Global styles updates:

*
[`dashboard/app/globals.css`](diffhunk://#diff-4e15958db8b4018955b23c01df1bd5f893495fb8eb6a73f263bd7d58220313efL78-R85):
Updated the color scheme for both light and dark modes to use Solarized
colors.
[[1]](diffhunk://#diff-4e15958db8b4018955b23c01df1bd5f893495fb8eb6a73f263bd7d58220313efL78-R85)
[[2]](diffhunk://#diff-4e15958db8b4018955b23c01df1bd5f893495fb8eb6a73f263bd7d58220313efL133-R134)

Component modifications:

*
[`dashboard/components/LanguagesBarChart.tsx`](diffhunk://#diff-cd6f2aff0f1a539be2a688025f517bb966e0e3b2f4399274e1e27e32b4e2222aL47-R71):
Changed the chart layout from vertical to horizontal, updated the chart
labels, and switched the `XAxis` and `YAxis` data keys accordingly.
*
[`dashboard/components/SideBar.tsx`](diffhunk://#diff-789a55c11f3c7300c5678f37f83847350e681afdd7621d5141a72296b473dd0cL20-R20):
Added a `variant` prop to the `Sidebar` component to support different
sidebar styles.