ci: Regularly Clear Caches (#460)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to clean
caches on a scheduled basis. The workflow is designed to run monthly and
can also be triggered manually.

New GitHub Actions workflow:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dR1-R24):
Added a new workflow named "Clean Caches" that runs on a monthly
schedule and can be triggered manually. The workflow includes steps to
check out the repository and delete all caches using the `gh cache
delete` command.