chore: Retry UI Tests (#463)
# Pull Request

## Description

This pull request includes changes to the testing setup to improve the
reliability of end-to-end tests by adding a new dependency and modifying
the test execution commands.

Testing improvements:

*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dR11):
Added `pytest-rerunfailures==15.0` to the list of dependencies.
*
[`tests/tests.just`](diffhunk://#diff-83e2fb4528694cfa62c213eefec53e6aeee0353030bf91ad4f830e69854b7918L12-R18):
Updated the `run` and `run-local` commands to include the `--reruns 2`
option, allowing failed tests to be retried up to two times.

Fixes #435