chore: Update GitHub Action Versions (#478)
# Pull Request

## Description

This pull request includes updates to the
`.github/workflows/code-checks.yml` file to ensure the use of the latest
versions of certain GitHub actions. The most important changes include
updating the versions of the `upload-sarif`, `init`, `analyze`, and
`sonarqube-scan` actions.

Updates to GitHub actions:

* Updated `upload-sarif` action to version `v3.28.9` in multiple steps
to ensure the latest features and fixes are used.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L75-R75)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L105-R105)
[[3]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L254-R254)
* Updated `init` and `analyze` actions to version `v3.28.9` to maintain
consistency with the `upload-sarif` action.
* Updated `sonarqube-scan` action to version `v5.0.0` to leverage the
latest improvements and capabilities.