chore: Update Dependabot Configuration (#492)
# Pull Request

## Description

This pull request modifies the `.github/dependabot.yml` file to change
the update schedule for various dependencies from weekly to daily. This
ensures that dependencies are checked and updated more frequently, which
can help in maintaining up-to-date and secure dependencies.

Changes to update schedules:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L13-R13):
Changed the update interval for GitHub Actions dependencies from weekly
to daily.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L31-R30):
Changed the update interval for Python test dependencies from weekly to
daily.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L46-R44):
Changed the update interval for TypeScript dependencies from weekly to
daily.