chore: Remove Pytest BDD Test Dependency (#512)
# Pull Request

## Description

This pull request includes significant changes to the end-to-end tests
for the homepage and sitemap, as well as modifications to the
dependencies in the `pyproject.toml` file. The main focus is on
refactoring the tests to remove the use of `pytest-bdd` and simplifying
the test structure.

Refactoring end-to-end tests:

* Removed `homepage.feature` file and its associated step definitions
from `test_homepage.py`
[[1]](diffhunk://#diff-5f94767a8075490f34031c9c7a0656e59b6d6772eac20b57e969f99a8872564eL1-L4)
[[2]](diffhunk://#diff-742992d8e066bb34983abd597cc431b88b73f7177b4cec2b76fd9779d4646c0aL1-L18).
* Removed `sitemap.feature` file and its associated step definitions
from `test_sitemap.py`
[[1]](diffhunk://#diff-eace947763caea7c4b2a51590c7a6e0c446c38676e2912086d501e546d91cdbbL1-L12)
[[2]](diffhunk://#diff-6a27ed68cca1634d5b12218027d4d764418a5ac01008704ec2f42f0ac0bbb064L1-L121).
* Added new test functions `test_homepage`, `test_sitemap_index`,
`test_sitemap`, and `test_sitemap_urls` in `test_homepage.py` and
`test_sitemap.py` to replace the removed BDD-style tests
[[1]](diffhunk://#diff-05b4ab29c19ba99ffc33cc0a624eaafb52db2a05af0bf871067f13083a63d72fR1-R8)
[[2]](diffhunk://#diff-ba4856c27cd44d538a01cd1404d46a9484ddd19c8af461a306825aac238fc621R1-R66).

Dependency updates:

* Removed `pytest-bdd` dependency from `pyproject.toml` as it is no
longer used in the refactored tests.