chore: Update Playwright UI Test Job (#514)
# Pull Request

## Description

This pull request includes changes to streamline the setup of test
dependencies and improve the efficiency of the testing process. The most
important changes include the removal of Chrome setup steps and the
addition of Playwright dependency installation steps.

### Changes to test dependency setup:

*
[`.github/actions/setup-test-dependencies/action.yml`](diffhunk://#diff-ee1e9a84357cfd23d067400706d92b6bb9d8fd8744e81563d9858e8c344df144L9-R16):
Removed the setup steps for Chrome and added steps to install Playwright
dependencies.

### Changes to testing scripts:

*
[`tests/tests.just`](diffhunk://#diff-83e2fb4528694cfa62c213eefec53e6aeee0353030bf91ad4f830e69854b7918R20-R27):
Added new tasks to install Playwright dependencies, including a task for
installing minimum dependencies with Chromium.