ci: Update runner versions (#556)
# Pull Request

## Description

This pull request includes several updates to configuration files to
ensure the use of the latest versions and dependencies. The most
important changes include updating the operating system version for
GitHub Actions workflows, updating the minimum version for `lefthook`,
and updating development dependencies in `pyproject.toml`.

Updates to GitHub Actions workflows:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L87-R87):
Changed the `runs-on` parameter from `ubuntu-22.04` to `ubuntu-latest`
for the `run-python-code-checks` job.
*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3L52-R52):
Changed the `runs-on` parameter from `ubuntu-22.04` to `ubuntu-latest`
for the `ui-tests` job.

Updates to configuration files:

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` from `1.11.2` to `1.11.6`.
*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL15-R18):
Updated development dependencies `ruff` from `0.9.9` to `0.11.2` and
`zizmor` from `1.4.1` to `1.5.2`. Updated `required-version` for
`tool.uv` from `0.6.9` to `0.6.11`.