chore: Remove Release Please (#566)
# Pull Request

## Description

This pull request removes the configuration and workflow files related
to the "Release Please" tool, as well as associated labels in the
repository. These changes likely indicate a decision to discontinue the
use of "Release Please" for managing releases.

### Removal of "Release Please" configurations:

*
[`.github/release-please/release-please-config.json`](diffhunk://#diff-f66a00044292c0645348bedeb8c0bf060e429f654585acb2ac641055cdc02f98L1-L44):
Deleted the configuration file for "Release Please," which included
changelog sections, package settings, and other release-related options.
*
[`.github/release-please/.release-please-manifest.json`](diffhunk://#diff-fe994ca589f1bf66bad114e09cd6c73cd858173e5ac917dce94b7af776984392L1-L3):
Removed the manifest file that tracked the release version for the
repository.
*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L1-L27):
Deleted the GitHub Actions workflow for automating releases using
"Release Please."

### Cleanup of related labels:

*
[`.github/other-configurations/labels.yml`](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dL54-L69):
Removed labels associated with "Release Please," such as `autorelease:
pending`, `autorelease: tagged`, and others.