docs: Add project `README.md` (#574)
# Pull Request

## Description

This pull request adds a new `README.md` file to provide documentation
for the project. The file includes an overview of the project,
contribution guidelines, a code of conduct, and licensing information.

Documentation updates:

* Added a `README.md` file with sections for project description, table
of contents, contributing guidelines, code of conduct, and license
details. This improves the accessibility and clarity of the project's
purpose and guidelines.