chore: Remove CodeQL Configuration (#578)
# Pull Request

## Description

This pull request removes a custom CodeQL configuration file and updates
the workflow to no longer reference it. These changes simplify the
CodeQL analysis setup in the repository.

### CodeQL configuration removal:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beL1-L3):
Deleted the query filter for excluding the `actions/unpinned-tag` query.

### Workflow update:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L194):
Removed the `config-file` parameter pointing to the deleted CodeQL
configuration file, ensuring the default configuration is used.