ci: Bump setup-uv to V6 (#587)
# Pull Request

## Description

This pull request updates the version of the `setup-uv` action used in
the repository to ensure compatibility with the latest features and
improvements. The changes affect two files:
`.github/actions/setup-test-dependencies/action.yml` and
`.github/workflows/code-checks.yml`.

Version update for `setup-uv` action:

*
[`.github/actions/setup-test-dependencies/action.yml`](diffhunk://#diff-ee1e9a84357cfd23d067400706d92b6bb9d8fd8744e81563d9858e8c344df144L10-R10):
Updated the `setup-uv` action from version `v5.4.2` to `v6.0.0` in the
"Install Python and UV" step.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L168-R168):
Updated the `setup-uv` action from version `v5.4.2` to `v6.0.0` in the
"Install the latest version of uv" step.