test: Add Repositories Content Tests (#625)
# Pull Request

## Description

This pull request introduces a new test for the homepage content and
refines an existing sitemap test by removing unnecessary logging. The
changes focus on improving test coverage and simplifying the codebase.

### Test coverage improvements:

*
[`tests/ui/test_homepage.py`](diffhunk://#diff-94f77bfc74fac1bb9b472a5379a0e22748abd216985b75fab2e5d0d52475b4d9R1-R61):
Added a new test, `test_homepage_content`, to validate the homepage
content, including the presence of a header, repository links, and their
attributes. This improves the verification of the user interface.

### Codebase simplification:

*
[`tests/ui/test_sitemap.py`](diffhunk://#diff-47de1e794455000e0e8a14d6a80753ee490acb471a6d040342eda69ac45293a8L63):
Removed a debug logging statement from the `test_sitemap_urls` function
to streamline the test and reduce unnecessary output.