test: Update homepage tests to repositories tests (#627)
# Pull Request

## Description

This pull request updates the UI test suite to reflect a shift in focus
from the homepage to the repositories page. The most important changes
include renaming a test file, updating test functions and variables, and
introducing a new URL for the repositories page.

### Updates to the UI test suite:

* Renamed the test file `tests/ui/test_homepage.py` to
`tests/ui/test_repositories.py` and updated its content to focus on
repositories instead of the homepage. This includes changes to test
function names, docstrings, and the URL being tested.

* Added a new variable `REPOSITORIES_URL` in
`tests/ui/utils/variables.py` to construct the repositories page URL
dynamically based on the `PROJECT_URL`.