chore: Set Test Python Version (#650)
# Pull Request

## Description

This pull request updates the Python environment requirements and
development dependencies in the `tests/pyproject.toml` file. The most
important changes include modifying the Python version specification and
updating the required version of a development tool.

### Updates to Python environment requirements:
* Changed the `requires-python` specification from `>=3.13` to `~=3.13`,
aligning with semantic versioning to allow for compatible minor updates.

### Updates to development dependencies:
* Updated the `required-version` of the `uv` tool from `0.7.3` to
`0.7.5` to ensure compatibility with the latest features or fixes.