test: Fix UI Tests (#662)
# Pull Request

## Description


This pull request updates a test case in `tests/ui/test_repositories.py`
to reflect a change in the expected number of repository links on the
page.

Test case update:

*
[`tests/ui/test_repositories.py`](diffhunk://#diff-18684b3b8dbb8b4785694160f807327ff53dd63b4400dbad6159d04e8fff675aL25-R25):
Modified the assertion in `test_repositories_content` to expect 32
repository links instead of 31, likely due to the addition of a new
repository link.