chore: Update Dependabot Groups Configuration (#689)
# Pull Request

## Description

This pull request includes updates to dependency configuration files and
a minor change to the Dependabot configuration. The changes primarily
focus on refining dependency management and ensuring consistency in
version specifications.

### Dependency configuration updates:

*
[`dashboard/package-lock.json`](diffhunk://#diff-ed953a47c378f3fc19c8e0152346ac0ffa9162d0ffacdbae96c4d9ffbcf479a6L9-L10):
Removed the `license` field and the `version` field from the root
package metadata. These fields are unnecessary in this context.
*
[`dashboard/package-lock.json`](diffhunk://#diff-ed953a47c378f3fc19c8e0152346ac0ffa9162d0ffacdbae96c4d9ffbcf479a6L39-R38):
Updated `@eslint/eslintrc` and `@eslint/js` dependencies to use exact
version specifications instead of version ranges for better consistency
and reproducibility.

### Dependabot configuration update:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L59-R59):
Adjusted the exclusion pattern for dependencies, replacing `"nextjs"`
with `"next"` to align with the correct naming convention.