chore: Fix UI Tests (#723)
# Pull Request

## Description

This pull request includes updates to dependencies and test cases in the
project. The most important changes involve upgrading development
dependencies, updating a test assertion to reflect an additional
repository, and adding a new repository name to the test data.

### Dependency updates:
*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL15-R18):
Upgraded `ruff` to version `0.12.2`, `ty` to version `0.0.1a14`, and
`uv`'s required version to `0.7.19`.

### Test updates:
*
[`tests/ui/test_repositories.py`](diffhunk://#diff-18684b3b8dbb8b4785694160f807327ff53dd63b4400dbad6159d04e8fff675aL25-R25):
Updated the test assertion in `test_repositories_content` to expect 33
repository links instead of 32.
*
[`tests/ui/test_repositories.py`](diffhunk://#diff-18684b3b8dbb8b4785694160f807327ff53dd63b4400dbad6159d04e8fff675aR37):
Added `"gitdash"` to the list of repository names in the
`test_repositories_content` test.