chore: Add actionlint Git Hook (#752)
# Pull Request

## Description

This pull request introduces a new `actionlint-check` task to the
development workflow, ensuring GitHub Actions workflows are validated
for best practices. The changes include updates to both the `Justfile`
and the `lefthook.yml` configuration.

### Development Workflow Enhancements:

* **Added `actionlint-check` task to `Justfile`:** This new task runs
the `actionlint` tool to validate GitHub Actions workflows. (`Justfile`,
[JustfileR56-R63](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR56-R63))
* **Integrated `actionlint-check` into pre-commit hooks:** Updated
`lefthook.yml` to include the `actionlint-check` task as part of the
pre-commit checks. (`lefthook.yml`,
[lefthook.ymlR23-R24](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR23-R24))