build: Remove Repositories Page (#766)
# Pull Request

## Description

This pull request removes the "Repositories" page and its associated
tests, as well as disables the TypeScript unit test workflow in the CI
pipeline. The changes streamline the codebase by eliminating unused
components and tests, and temporarily disable a workflow for further
review or updates.

### Removal of the "Repositories" page:

*
[`dashboard/app/repositories/page.tsx`](diffhunk://#diff-a4ccef6424b97b0ba92144755234e88a172fc73da6bad5204fde60651a96afa3L1-L34):
Deleted the entire "Repositories" page component, which included
functionality to display a list of repositories with links.
*
[`dashboard/components/SideBar.tsx`](diffhunk://#diff-789a55c11f3c7300c5678f37f83847350e681afdd7621d5141a72296b473dd0cL47-R50):
Updated the sidebar to replace the "Repositories" link with a static
`div` element, removing the hyperlink functionality.

### Removal of associated tests:

*
[`dashboard/app/__tests__/page.spec.tsx`](diffhunk://#diff-694f8c76deb82890cdb08dcec7b90eb1575d51cf79f15862257994d5081921e7L1-L13):
Removed the test file for the "Repositories" page, which included a test
to verify the rendering of the page.
*
[`dashboard/app/__tests__/__snapshots__/page.spec.tsx.snap`](diffhunk://#diff-78cc71780ea0095b5a694235f1d2ec04873ae0b72e999e94f089a7b0d6a986b3L1-L3):
Deleted the Jest snapshot file associated with the "Repositories" page
tests.

### Workflow changes:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L202-R223):
Commented out the `typescript-unit-tests` job in the CI pipeline,
effectively disabling it for now. This job included steps for setting up
dependencies, running unit tests, and performing a SonarCloud scan.