chore: Add Test Plan (#776)
# Pull Request

## Description

This pull request removes the Jest configuration and setup files from
the `dashboard` directory, likely as part of a cleanup or migration
effort.

### Removal of Jest configuration:

*
[`dashboard/jest.config.ts`](diffhunk://#diff-cd42b53283f5b867fd6ecb86064ba8ab41c776420160f887563d2470d2098dc9L1-L20):
Deleted the Jest configuration file, which included custom
configurations and integration with Next.js via `next/jest`.

### Removal of Jest setup:

*
[`dashboard/jest.setup.ts`](diffhunk://#diff-bd1c51eb8d79435a4760b607fe8b414276c8332f926f73db88ed8dc931159cbbL1-L3):
Deleted the Jest setup file, which had been used to polyfill
`ResizeObserver` in the test environment.