chore: Refactor OverviewPage to SummaryPage (#778)
# Pull Request

## Description

This pull request refactors the `OverviewPage` component in
`dashboard/app/page.tsx` to rename key identifiers for improved clarity
and consistency. The most important changes include renaming the
component, data variables, and imported modules to align with the new
naming convention.

### Refactoring for clarity and consistency:

* Renamed the `OverviewPage` component to `SummaryPage` and updated all
references to the component accordingly.
* Renamed the `OverviewKeyFacts` component to `SummaryKeyFacts` in the
import statement and updated its usage in the JSX.
[[1]](diffhunk://#diff-e1b87d8e32d6a6029b017f01e52506c46844e0d2a3a7fbd57decc879e5630d6fL3-R8)
[[2]](diffhunk://#diff-e1b87d8e32d6a6029b017f01e52506c46844e0d2a3a7fbd57decc879e5630d6fL79-R84)
* Renamed the `getOverviewData` function to `getSummaryData` in the
import statement and updated all references to the function.
* Updated variable names from `overviewData` to `summaryData` to reflect
the new naming convention.
[[1]](diffhunk://#diff-e1b87d8e32d6a6029b017f01e52506c46844e0d2a3a7fbd57decc879e5630d6fL3-R8)
[[2]](diffhunk://#diff-e1b87d8e32d6a6029b017f01e52506c46844e0d2a3a7fbd57decc879e5630d6fL24-R32)
[[3]](diffhunk://#diff-e1b87d8e32d6a6029b017f01e52506c46844e0d2a3a7fbd57decc879e5630d6fL41-R49)
[[4]](diffhunk://#diff-e1b87d8e32d6a6029b017f01e52506c46844e0d2a3a7fbd57decc879e5630d6fL79-R84)

These changes improve the code's readability and ensure consistency in
naming conventions throughout the file.