deps(github-actions): Update Reusable Workflows (#782)
# Pull Request

## Description

This pull request updates references to reusable workflows across
multiple GitHub Actions workflow files to use a newer version
(`v2025.08.04.01`) of the workflows. The updates ensure that the latest
improvements and fixes in the reusable workflows are applied.

### Workflow updates:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated the reusable workflow reference for cache cleaning to version
`v2025.08.04.01`.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L187-R187):
Updated the reusable workflow references for code checks and CodeQL
analysis to version `v2025.08.04.01`.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L187-R187)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L199-R199)
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL24-R24):
Updated the reusable workflow reference for pull request tasks to
version `v2025.08.04.01`.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L19-R19):
Updated the reusable workflow reference for syncing labels to version
`v2025.08.04.01`.