chore: Fix UI Tests (#831)
# Pull Request

## Description

This pull request updates the repository list used in the
`test_sidebar_repositories_list` test in `tests/ui/test_sidebar.py`. The
main change is the reordering and addition of several repository names
to ensure the sidebar test covers the current set of repositories.

Repository list updates:

* Added new repositories to the test list, including `coding-metrics`,
`exercism`, `remove-pr-bot-collaborators`, `test-coding-metrics`, and
others to reflect the latest repository set.
* Reordered some repository names and restored previously removed
entries (`projects`, `RepoOrchestrator`, `TechInsight`) to maintain
consistency and completeness in the test.