ci: Update Pull Request Tasks Permissions (#870)
# Pull Request

## Description

This pull request makes a minor adjustment to the permissions
configuration in the GitHub Actions workflow file
`.github/workflows/pull-request-tasks.yml`. The change moves the
`pull-requests: read` permission from the global workflow level to the
specific job that requires it, resulting in a more restrictive and
secure permissions setup.

- Workflow permissions refactor:
* Removed the global `pull-requests: read` permission and replaced it
with an empty permissions object at the workflow level, then added
`pull-requests: read` permission specifically to the `check-pr-title`
job. This limits the permission scope to only where it's needed.